break up user.go
